LayoutTests/imported/w3c/web-platform-tests/svg/types/scripted/SVGGraphicsElement.getBBox-02-expected.txt

XX

FAIL SVGGraphicsElement.prototype.getBBox on <tspan> assert_equals: x expected 100 but got 0
